+++ plugins/tracker/src/index.ts
@@ -112,6 +112,7 @@ export interface Issue extends AttachedDoc {
   subIssues: number
   blockedBy?: Ref<Issue>[]
   relatedIssue?: Ref<Issue>[]
+  parents: IssueParentInfo[]
 
   comments: number
   attachments?: number
@@ -122,6 +123,14 @@ export interface Issue extends AttachedDoc {
   dueDate: Timestamp | null
 
   rank: string
+}
+
+/**
+ * @public
+ */
+export interface IssueParentInfo {
+  parentId: Ref<Issue>
+  parentTitle: string
 }
 
 /**